Assignment – 2 (1 mark)
Create a new webpage that is enhanced with CSS. First create a new HTML page that includes a minimum of text (including headers and paragraph tags), images, and at least one table. This page may contain any information that you want - it can even use made up text from a text generator such as lipsum.com. Do not style in anyway with HTML, instead, create and link and external style sheet to this page that changes the format of some CSS properties. 
To complete this assignment, please complete all of the following requirements: 
 Create an HTML page with a header and body section that includes some text, at least one image, and at least one table (with as many rows and columns as you want) 
 Create an external CSS file that is properly linked to the webpage you've designed, that includes at least 5 different classes or IDs that modify elements of the linked webpage. 
 At least one of the classes or IDs must control layout using layers (Div tag), and the rest should modify properties such as fonts, colors, background-color, etc. If you want an extra challenge, research how to control some other CSS properties that you think would be useful when designing webpages and use it. Remember W3schools provides excellent information in this regard and the ability to practice in order to learn what specific CSS properties control.